frontend/api/requirements.txt
openai>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
pymupdf>=1.18.0